Mudanças na classe personagem e criação de classe Chalice
