+++ Src/Classes/chaliceImgLists.py
@@ -1,0 +1,12 @@
+
+ChaliceNormal = []
+for i in range(1, 5):
+    ChaliceNormal.append(f"../Img/Chalice/Test{i:0d}.png")
+    
+ChaliceTransitionToSpecial = []
+for i in range(1, 19):
+    ChaliceTransitionToSpecial.append(f"../Img/Chalice/Special/chalice_shmup_super_intro_{i:04d}.png")
+
+ChaliceSpecial = []
+for i in range(1, 5):
+    ChaliceSpecial.append(f"../Img/Chalice/Special/chalice_shmup_super_idle_straight_{i:04d}.png")
