+++ Src/main.py
@@ -1,11 +1,10 @@
 from tupy import*
 from Classes.armas import Arma, PedraSeno, FogoReto, PedraY3, PedraY5
-from Classes.personagem import Vilao_test_1, Heroi_test_1
 import pygame
 from pygame.locals import * 
 from Classes.cenario import *
 from HildaBerg.hildaBerg import *
-
+from Classes.chalice import Chalice
 pygame.init()
 
 e = Ceu(500, 280, 4)
@@ -24,6 +23,6 @@ pedray5 = Arma(1, PedraY5)
 #vi = Vilao_test_1(800, 300, ["../Img/blimp_idle_0001.png", "../Img/blimp_idle_0001.png" ],200, [pedra, fogo, pedray3, pedray5])
 
 vi = HildaBerg(700, 220)
-s = Heroi_test_1(["../Img/Test1.png", "../Img/Test2.png", "../Img/Test3.png", "../Img/Test4.png"], vi)
+s = Chalice(vi)
 
 run(globals())
